Updated according to comments from https://github.com/robotframework/robotframework/pull/1852
