+++ src/robot/running/keywords.py
@@ -297,52 +297,3 @@ class _ForItem(_BaseKeyword):
         self.status = status
         self.endtime = get_timestamp()
         self.elapsedtime = get_elapsed_time(self.starttime, self.endtime)
-
-def numberOfDigits(number):
-    digits = 0
-    import re
-    from locale import localeconv
-    convertedNumber = str(number)
-    dec_pt = localeconv()['decimal_point']
-    #split the string in 2 parts if necessary
-    list_of_strings = convertedNumber.split(dec_pt)
-    if len(list_of_strings) == 1:
-        #there was no decimal point
-        digits = 0
-    elif len(list_of_strings) == 2:
-        #regular float number
-        digits = len(list_of_strings[1])
-    else:
-        #invalid input
-        raise ValueError, "input is not a number"
-    return digits    
-
-def frange(*args):
-    result = []
-    #find out the arguments
-    if len(args) == 1:
-        start = 0
-        stop = eval(args[0])
-        step = 1
-    elif len(args) == 2:
-        start = eval(args[0])
-        stop = eval(args[1])
-        step = 1
-    elif len(args) == 3:
-        start = eval(args[0])
-        stop = eval(args[1])
-        step = eval(args[2])
-    else:
-        raise ValueError, "invalid number of arguments"
-    #find out maximum number of decimals
-    powerOf10 = max(numberOfDigits(start),numberOfDigits(stop),numberOfDigits(step))    
-    if powerOf10 == 0:
-        result = range(start, stop, step)
-    else:
-        #multiply each argument to avoid rounding errors if necessary
-        factor = pow(10,powerOf10)    
-        result = [x / float(factor) for x in range(int(start*factor),int(stop*factor),int(step*factor))]
-    return result
-
-def num(s):
-    return float(s) if '.' in str(s) else int(s)
